+++ src/nvim/drawscreen.c
@@ -1643,12 +1643,6 @@ static void win_update(win_T *wp)
         top_end = 1;
       }
     }
-
-    // When line numbers are displayed need to redraw all lines below
-    // inserted/deleted lines.
-    if (mod_top != 0 && buf->b_mod_xlines != 0 && wp->w_p_nu) {
-      mod_bot = MAXLNUM;
-    }
   }
 
   wp->w_redraw_top = 0;  // reset for next time
@@ -2326,9 +2320,12 @@ static void win_update(win_T *wp)
       idx++;
       lnum += foldinfo.fi_lines + 1;
     } else {
-      if (wp->w_p_rnu && wp->w_last_cursor_lnum_rnu != wp->w_cursor.lnum) {
-        // 'relativenumber' set and cursor moved vertically: The
-        // text doesn't need to be drawn, but the number column does.
+      // If:
+      // - 'number' is set and below inserted/deleted lines, or
+      // - 'relativenumber' is set and cursor moved vertically,
+      // the text doesn't need to be redrawn, but the number column does.
+      if ((wp->w_p_nu && mod_top != 0 && lnum >= mod_bot && buf->b_mod_xlines != 0)
+          || (wp->w_p_rnu && wp->w_last_cursor_lnum_rnu != wp->w_cursor.lnum)) {
         foldinfo_T info = wp->w_p_cul && lnum == wp->w_cursor.lnum
                           ? cursorline_fi : fold_info(wp, lnum);
         win_line(wp, lnum, srow, wp->w_grid.rows, wp->w_lines[idx].wl_size, &spv, info);
