+++ test/old/testdir/test_number.vim
@@ -146,8 +146,7 @@ func Test_number_with_linewrap1()
   call s:close_windows()
 endfunc
 
-" Pending: https://groups.google.com/forum/#!topic/vim_dev/tzNKP7EDWYI
-func XTest_number_with_linewrap2()
+func Test_number_with_linewrap2()
   call s:test_windows(3, 20)
   normal! 61ia
   setl number wrap
@@ -164,8 +163,7 @@ func XTest_number_with_linewrap2()
   call s:close_windows()
 endfunc
 
-" Pending: https://groups.google.com/forum/#!topic/vim_dev/tzNKP7EDWYI
-func XTest_number_with_linewrap3()
+func Test_number_with_linewrap3()
   call s:test_windows(4, 20)
   normal! 81ia
   setl number wrap
@@ -174,7 +172,7 @@ func XTest_number_with_linewrap3()
   call s:validate_cursor()
   let lines = s:screen_lines(1, 4)
   let expect = [
-\ "aaaaaaaa",
+\ "<<<aaaaa",
 \ "aaaaaaaa",
 \ "aaaaaaaa",
 \ "a       ",
@@ -345,6 +343,31 @@ func Test_relativenumber_callback()
   let buf = RunVimInTerminal('-S Xrnu_timer', #{rows: 8})
   call TermWait(buf, 310)
   call VerifyScreenDump(buf, 'Test_relativenumber_callback_1', {})
+
+  call StopVimInTerminal(buf)
+endfunc
+
+" Test that line numbers below inserted/deleted lines are updated.
+func Test_number_insert_delete_lines()
+  CheckScreendump
+
+  let lines =<< trim END
+      call setline(1, range(1, 7))
+      set number
+      call cursor(2, 1)
+  END
+  call writefile(lines, 'Xnumber_insert_delete_lines', 'D')
+
+  let buf = RunVimInTerminal('-S Xnumber_insert_delete_lines', #{rows: 8})
+  call VerifyScreenDump(buf, 'Test_number_insert_delete_lines_1', {})
+  call term_sendkeys(buf, "dd")
+  call VerifyScreenDump(buf, 'Test_number_insert_delete_lines_2', {})
+  call term_sendkeys(buf, "P")
+  call VerifyScreenDump(buf, 'Test_number_insert_delete_lines_1', {})
+  call term_sendkeys(buf, "2dd")
+  call VerifyScreenDump(buf, 'Test_number_insert_delete_lines_3', {})
+  call term_sendkeys(buf, "P")
+  call VerifyScreenDump(buf, 'Test_number_insert_delete_lines_1', {})
 
   call StopVimInTerminal(buf)
 endfunc
